Finished ParallelPrefixScan.comp (initial pass), plus retrofitting the ShaderStorage to be able to create composite shaders.
The composite shader assembly will help with the compute shaders because
it will allow me to hard-code buffer locations and intended sizes and
use a "constants" C-style file with #ifndef/#define/#endif blocks, and
then use that file (or multiple such files) in both compute shaders and
as a C-kosher include file in my compute controllers.  This should allow
me to reduce coupling between my SSBOs and misc. compute shaders.  But I
have to be careful with the file order.