Woot!  I now have a basic parallel prefix sum working over a size 16 array of 1s and 0s.
It wasn't working properly for a while, and then I found that OpenGL
won't throw a fit if you do an array overrun.  So I reexamined how big
my SSBO needed to be (1 work group for all data, even if that data only
fills part of the SSBO and the rest needs to be 0s) and made sure that
it was all initialized to 0s, and then it worked.  Interesting.  I
thought that my glMapBufferRange(...) wasn't set right.

Anyway, on to larger data sets!