now handling 100,000 items
It took a bit of tweaking and bug fixing, but I am now calculating the
prefix sums on the GPU for non-power-of-2 data sizes.  The algorithm
relies on a binary tree and thus on data sets that are a power of 2, but
I tweaked the SSBO and the uMaxThreads uniform in
ParticleSort::Sort(...) to account for this by making sure that all data
was part of a work group and then filling in the extra space with 0s.
The 0s will not affect the prefix sums.