Adding and removing files, does NOT work
I added a number of new compute shaders and "compute headers", if you
will, for buffers that will be used in those shaders.  I deleted a
number of shaders that aren't being used in this demo, and I am about to
do similar cleanup with the existing SSBOs that were meant for other
demos.