AAAAND IT WORKS!!
The problem was my calculation of the number of work groups that were
being dispatched.  I did this: ((num data entries) / (items per work
group + 1)) + 1.  The inner +1 was to deal with the case of data size
being equally divisible by the number of work groups, but when the data
size exceeded 512*512 (262144), then that +1 caused division by 512+1 to
result in a quotient 1 smaller than it should have been.  The problem
repeated at 512*1024 and 512*1536 and continued on.  The prefix scan was
dividing the data size by 1024 (2 * work group size), but it would have
had the same problem once the data size exceded 1024x1024.

Fixed by making an explicit condition check for the "is equally
divisible by" condition and handling the work group calculation as
necessary.