Made optimizations, fixed Stopwatch's LapTime method.  Now need to profile and clean up.
Now using a single IntermediateDataBuffer instead of double size instead
of "first" and "second" IntermediataData buffers.  This eliminated some
conditions in GetBitForPrefixScan.comp and in SortIntermediateData.comp.

No longer flipping the bits in GetBitForPrefixSum.comp.  I was following
the lead of the GPU Gems 3 chapter 39 article on a radix sorting
algorithm.  I don't know why they flipped the bits.  It made sense to
someone to calculate the prefix sum of 0s instead of 1s.