Fixed a "read index" bug in GetBitForPrefixScan.comp.  Need cleanup.
The problem was that I was using the same index value that read from the
IntermediateDataBuffer, which is now twice the size of the prefix sum
array because it is a single read/write buffer instead of separate
buffers, as the write index into the prefix sum array.  I still don't
know why the data memory was doing what it did.